Add files via upload
Fix some bugs in between-group comparisons of categorical variables. It is only a temporal solution -- now, different levels of a categorical variable could only be represented by characters, such as "Control" and "Patient" for the variable "Diagnosis", and "Male" and "Female" for the variable "Sex". In the final solution, both characters (such as "Male" and "Female") and numeric (such as 0 and 1) could represent the levels of a categorical variable.